refactor: screen 리팩토링
